requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.1
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
sweetify==2.3.1
typing_extensions==4.12.2
